+++ packages/types/src/interfaces/extrinsic/v1/Extrinsic.ts
@@ -1,10 +1,19 @@
+import {cvmToAddress} from "@cennznet/types/utils";
 import { GenericExtrinsic, UInt } from '@polkadot/types';
 import { SignatureOptions } from '@cennznet/types/interfaces/extrinsic/types';
-import { AnyU8a, ExtrinsicPayloadValue, IKeyringPair, Registry } from '@polkadot/types/types';
+import {
+  AnyU8a,
+  Callback,
+  ExtrinsicPayloadValue,
+  IKeyringPair,
+  ISubmittableResult,
+  Registry
+} from '@polkadot/types/types';
 import { Address, Call } from '@polkadot/types/interfaces/runtime';
 import { ExtrinsicValueV4 } from '@polkadot/types/extrinsic/v4/Extrinsic';
 import { Api } from '@cennznet/api';
 import { EstimateFeeParams, PaymentOptions } from '@cennznet/api/derives/types';
+import {base64Encode} from "@polkadot/util-crypto";
 import {Constructor, HexString} from '@polkadot/util/types';
 
 export default class CENNZnetExtrinsic extends GenericExtrinsic{
@@ -77,6 +86,21 @@ export default class CENNZnetExtrinsic extends GenericExtrinsic{
           console.error(err)
           return this;
       })
+    return this;
+  }
+
+  /**
+   * @description sign the extrinsic via any ethereum wallet
+   */
+  async signViaEthWallet(ethAddress: string, api: Api, ethereum: any, optionalStatusCb?: Callback<ISubmittableResult>): Promise<this> {
+    const cennznetAddress = cvmToAddress(ethAddress);
+    const nonce = await api.rpc.system.accountNextIndex(cennznetAddress);
+    const payload = this.registry.createType('EthWalletCall', { call: this, nonce }).toHex();
+    const encodedPayload = `data:application/octet-stream;base64,${base64Encode(payload)}`;
+    // Request signature from ethereum wallet
+    const signature = await ethereum.request({ method: 'personal_sign', params: [encodedPayload, ethAddress] });
+    // Broadcast the tx to CENNZnet
+    await api.tx.ethWallet.call(this, ethAddress, signature).send(optionalStatusCb);
     return this;
   }
 }